sadeem-rag-chatbot/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
chromadb==0.4.24
sentence-transformers==3.0.1
transformers==4.41.2
torch==2.3.1
google-generativeai==0.5.0
google-ai-generativelanguage==0.6.1
pandas==2.2.0
numpy==1.26.0
python-dateutil==2.8.2